pytest/requirements.txt
selene
webdriver_manager
pytest
pytest-selenium
